Refactored orionldPatchAttribute.cpp, put kjAttributeKeyValueAmend in its own module
